+++ lib/generators/rails_template18f/sidekiq/templates/config/initializers/redis.rb
@@ -1,15 +1,14 @@
 # frozen_string_literal: true
 
-if ENV["VCAP_SERVICES"].present?
-  redis_config = JSON.parse(ENV["VCAP_SERVICES"])["aws-elasticache-redis"]&.first
-  return if redis_config.nil?
-  redis_url = redis_config["credentials"]["uri"]
+Rails.application.config.to_prepare do
+  redis_url = CloudGovConfig.dig "aws-elasticache-redis", "credentials", "uri"
+  if redis_url.present?
+    Sidekiq.configure_server do |config|
+      config.redis = {url: redis_url, ssl: true}
+    end
 
-  Sidekiq.configure_server do |config|
-    config.redis = {url: redis_url, ssl: true}
-  end
-
-  Sidekiq.configure_client do |config|
-    config.redis = {url: redis_url, ssl: true}
+    Sidekiq.configure_client do |config|
+      config.redis = {url: redis_url, ssl: true}
+    end
   end
 end